refactor($core): extract internal plugin 'override-css'
